Move common fakes to sharedTest dir and introduce test modules
