+++ gnd/src/sharedTest/java/com/google/android/gnd/persistence/remote/FakeRemoteStorageManager.java
