+++ gnd/src/sharedTest/java/com/google/android/gnd/persistence/uuid/FakeUuidGenerator.java
