+++ gnd/src/sharedTest/java/com/google/android/gnd/system/auth/FakeAuthenticationManager.java
